+++ src/componenets/AdminPanel/treatment/GetTreatment.jsx
@@ -20,9 +20,9 @@ const GetTreatment = () => {
 
     useEffect(() => {
         if (Array.isArray(allTreatmentData)) {
-          setShowData(allTreatmentData?.slice(firstNumber, secondNumber));
+            setShowData(allTreatmentData?.slice(firstNumber, secondNumber));
         }
-      }, [allTreatmentData, firstNumber, secondNumber]);
+    }, [allTreatmentData, firstNumber, secondNumber]);
 
     const handlePrevious = (e) => {
         e.preventDefault();
@@ -36,9 +36,12 @@ const GetTreatment = () => {
         setSecondNumber(secondNumber + 15);
     }
 
-    const handleTreatmentDelete = (id) => {
-        dispatch(deleteTreatment(id));
-        dispatch(getAllTreatment());
+    const handleTreatmentDelete = async (id) => {
+        const result = window.confirm("Are you sure want to delete?");
+        if (result) {
+            dispatch(deleteTreatment(id));
+            dispatch(getAllTreatment());
+        }
     }
 
     const handlePainAreaUpdate = (Values) => {
@@ -72,7 +75,7 @@ const GetTreatment = () => {
                                     <td>{item?.titleEs}</td>
                                     <td>{item?.treatmentLevel}</td>
                                     <td>{item?.duration}</td>
-                                    <td><video style={{width: "75px", height: "75px", padding: "0px"}} src={item?.treatmentUrl}></video></td>
+                                    <td><video style={{ width: "75px", height: "75px", padding: "0px" }} src={item?.treatmentUrl}></video></td>
                                     <td onClick={() => handlePainAreaUpdate(item)}><i className='fa fa-pencil-square'></i></td>
                                     <td onClick={() => handleTreatmentDelete(item._id)}><i className='fa fa-trash'></i></td>
                                 </tr>
